Moved all of the Entity class construction code into entity.js
